Improved docs, moved skin to an extra file, improved utils.py
