--- docs/minepi.rst
@@ -1,15 +1,0 @@
-*************
-API Reference
-*************
-.. toctree::
-    :maxdepth: 2
-    :caption: Contents
-
-.. autoclass:: minepi.Player
-    :members:
-
-.. autoclass:: minepi.Skin
-    :members:
-
-.. automodule:: minepi.utils
-    :members:
